+++ asv/results.py
@@ -16,6 +16,7 @@ import datetime
 import six
 from six.moves import zip as izip
 
+from asv.config import flatten_env_matrix, normalize_env_matrix
 from . import environment
 from .console import log
 from .machine import Machine
@@ -192,7 +193,14 @@ class Results(object):
     """
     api_version = 1
 
-    def __init__(self, params, requirements, commit_hash, date, python, env_name):
+    def __init__(self,
+                 params,
+                 requirements,
+                 commit_hash,
+                 date,
+                 python,
+                 env_name,
+                 env_vars_combination):
         """
         Parameters
         ----------
@@ -215,6 +223,9 @@ class Results(object):
 
         env_name : str
             Environment name
+
+        env_vars_combination: dict
+            Combination of additional environment variables
         """
         self._params = params
         self._requirements = requirements
@@ -230,6 +241,7 @@ class Results(object):
         self._started_at = {}
         self._ended_at = {}
         self._benchmark_version = {}
+        self._env_vars_combination = env_vars_combination
 
         # Note: stderr and errcode are not saved to files
         self._stderr = {}
@@ -243,7 +255,7 @@ class Results(object):
 
     @classmethod
     def unnamed(cls):
-        return cls({}, {}, None, None, None, None)
+        return cls({}, {}, None, None, None, None, {})
 
     @property
     def commit_hash(self):
@@ -256,6 +268,10 @@ class Results(object):
     @property
     def params(self):
         return self._params
+
+    @property
+    def env_vars_combination(self):
+        return self._env_vars_combination
 
     @property
     def started_at(self):
@@ -584,6 +600,9 @@ class Results(object):
 
         data = {
             'results': results,
+            'env_vars_combination': normalize_env_matrix(
+                self._env_vars_combination
+            ),
             'params': self._params,
             'requirements': self._requirements,
             'commit_hash': self._commit_hash,
@@ -610,6 +629,7 @@ class Results(object):
         if os.path.isfile(path):
             old = self.load(path)
             for dict_name in ('_results', '_samples', '_stats',
+                              '_env_vars_combination',
                               '_benchmark_params', '_profiles', '_started_at',
                               '_ended_at', '_benchmark_version'):
                 setattr(self, dict_name, getattr(old, dict_name))
@@ -628,8 +648,12 @@ class Results(object):
 
         """
         d = util.load_json(path, cls.api_version)
+        d.setdefault('env_vars_combination', {})
 
         try:
+            env_vars_combination = flatten_env_matrix(
+                d['env_vars_combination']
+            )
             obj = cls(
                 d['params'],
                 d['requirements'],
@@ -637,7 +661,11 @@ class Results(object):
                 d['date'],
                 d['python'],
                 d.get('env_name',
-                      environment.get_env_name('', d['python'], d['requirements']))
+                      environment.get_env_name('',
+                                               d['python'],
+                                               d['requirements'],
+                                               env_vars_combination)),
+                env_vars_combination,
             )
 
             obj._results = {}